Expanding the APIs for transports and callbacks
This commit adds support to get the request objects of whatever
server used for the transports. This information is useful when
building tools over socket.io-erlang where smarter dispatching
and access to queries, headers, etc. is required.

This also adds the support for custom arguments in the default
http handler in order to build sessions or routing information
for more complex behaviours to be built over them.

All changes are backward compatible.