Warning messages on configuration loading failure
The current behaviour of the application when the configuration value
for heartbeats (and a delay for closing files in xhr) was to silently
ignore the issue and substitute it by a default value.

This patch adds warning messages going through the standard error_logger
in order to let the user of socket.io-erlang know something went wrong.

The default heartbeat value from websockets was also changed from
infinity to 10 seconds, the same as in the app file we currently have.
Setting it to infinity had as a consequence that the client code would
always disconnect from the protocol, without closing the websocket. This
would result in a particular tab of a browser holding dozens of live
websockets that can receive data (and look alive to the server) but not
send any from all of them but the last one.