Fixing timers in websockets
The current way of doing timers in websockets is fickle and bound to
error when it is reset on all messages we get. Just by sending a process
random messages, the timers get reset all the time.

This fix adds a different way of handling timers by using the erlang
module's start_timer/3 and cancel_timer/1 to get a safer
timeout/heartbeat behaviour.