Reworking the timers in transport_polling
The current way of handling timeouts is risky with regards to user calls
able to mess with the timeouts and get things out of sync with the
client side.

This fix replaces the timeouts coming from the gen_server behaviour with
the erlang timer BIFs, likely safer and independent of the messages
received.

The timeouts having to do with a connection being closed are still
handled the old way.