Fixing more timers
More timers and heartbeats fixed, including htmlfile and xhr multipart.
Nothing in the tests seemed to have an influence on the htmlfile
transport. Further tests ought to be added in time.

Note that as for the other transport, the timeout on a disconnection is
still not a BIF, compared to the heartbeats which now use the Erlang
timer functions.