Fixing the timeout events on transports.
We fix the timer issues having to do with erlang BIF timers conflicting
with the connection timeouts occurring after a closed socket.

Any client whose connection is closed at the point where the 'dead
connection timeout' is longer than the polling/heartbeat timer's time
left will have conflicting timers where the polling/heartbeat resets the
dead connection timeout.

Given the delays between the 'dead connection timeout' and the regular
polling/heartbeat timeout can be very close (and are set that way by
default), this conflict happens fairly often.

Because we are not exactly sure what makes it valid to cancel the dead
connection timer, this fix makes it so both timers are handled the same
way, replicating the old behaviour present before the recent fixes to
the timers when it comes to these connections.