Fixing build problems (issue #61)
This commit upgrades the current agnerized version being used by
the Makefile. This prompted changes to the rebar config.

The true build problems were caused by the merge #63, which added
a fix to the -include_lib path to better work with release.
However, this introduced a problem when compiling with rebar
given rebar calls 'erlc', which ignores all ERL_LIBS options
and thus will only support the default system libs (apparently).

This commit adds a macro definition to rebar.config and a
preprocessor switch to the socketio_listener module (which had
the include_lib changed) so that when compiled from rebar, we
use the old code that worked, and when going without rebar, we
use the fix that existed.

Hopefully that will make everyone happy.