Allowing multiple instances of socket.io connects
This is done by using dynamic atoms (a bad thing) based on
the number of ports you have a connection on.

A good design would have required a substantial rewrite of
the application. However, to simply get things working in
a decent way, although not perfect in terms of patterns,
I decided to go this way. It is worth the time for an
application in maintenance -- no technical debt on this
one, unless someone else wants to take control.